增加搜索记录控制器(根据条件请求数据、请求搜索类型、请求图表前30天数据)文件与功能

+++ rcc-api-commons/src/main/java/com/aenlly/rcc/entity/PointsLog.java
@@ -37,7 +37,7 @@ public class PointsLog implements Serializable {
 
   @ApiModelProperty(value = "增加/消耗的积分")
   @TableField("number")
-  private Integer number;
+  private Long number;
 
   @ApiModelProperty(value = "积分记录类型，1：增加/-1：减少")
   @TableField("type")
